extraction de main.java du package Personnages. YOLO
